pyproject.toml
[tool.poetry]
name = "roi2nix"
version = "0.1.0"
description = ""
authors = ["David Parker <davidparker@flywheel.io>"]

[tool.poetry.dependencies]
python = "~3.8"
flywheel-sdk = "15.8.0"
flywheel-gear-toolkit = "0.3.1"
nibabel = "3.2.1"
numpy = "1.22.1"
scikit-image = "0.18.1"
pydicom = "2.3.0"
dcmpi = "0.0.2.8"
pynrrd = "0.4.2"
scipy = "1.7.3"
python-gdcm = "^3.0.10"
pylibjpeg = "1.4.0"
pylibjpeg-libjpeg = "1.3.0"
dicom2nifti = "~2.3.0"
fw-file = {version = "^1.3.3", extras = ["all"]}

[tool.poetry.dev-dependencies]
pytest = "7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

